fix spectator

+++ ocgcore/ocgapi.cpp
@@ -298,7 +298,7 @@ extern "C" DECL_DLLEXPORT int32 query_field_info(ptr pduel, byte* buf, bool isSw
 			if(pcard) {
 				*buf++ = 1;
 				*buf++ = pcard->current.position;
-				*buf++ = pcard->xyz_materials.size();
+				*buf++ = (uint8) pcard->xyz_materials.size();
 			} else {
 				*buf++ = 0;
 			}
@@ -312,14 +312,14 @@ extern "C" DECL_DLLEXPORT int32 query_field_info(ptr pduel, byte* buf, bool isSw
 				*buf++ = 0;
 			}
 		}
-		*buf++ = ptduel->game_field->player[playerid].list_main.size();
-		*buf++ = ptduel->game_field->player[playerid].list_hand.size();
-		*buf++ = ptduel->game_field->player[playerid].list_grave.size();
-		*buf++ = ptduel->game_field->player[playerid].list_remove.size();
-		*buf++ = ptduel->game_field->player[playerid].list_extra.size();
-		*buf++ = ptduel->game_field->player[playerid].extra_p_count;
+		*buf++ = (uint8) ptduel->game_field->player[playerid].list_main.size();
+		*buf++ = (uint8) ptduel->game_field->player[playerid].list_hand.size();
+		*buf++ = (uint8) ptduel->game_field->player[playerid].list_grave.size();
+		*buf++ = (uint8) ptduel->game_field->player[playerid].list_remove.size();
+		*buf++ = (uint8) ptduel->game_field->player[playerid].list_extra.size();
+		*buf++ = (uint8) ptduel->game_field->player[playerid].extra_p_count;
 	}
-	*buf++ = ptduel->game_field->core.current_chain.size();
+	*buf++ = (uint8) ptduel->game_field->core.current_chain.size();
 	for(auto chit = ptduel->game_field->core.current_chain.begin(); chit != ptduel->game_field->core.current_chain.end(); ++chit) {
 		effect* peffect = chit->triggering_effect;
 		*((int*)(buf)) = peffect->handler->data.code;
